Add instructions on using git replace
Something the Rust Prehistory repository can be good for is seeing the
full evolution of test cases over time, to see how the language evolved.
Add instructions to the README on how to use "git replace" to achieve a
full, integrated history.  Useful for people who may not know about "git
replace", or those who do who don't want to be bothered to figure out
the commits that need to be replaced.